Move decode/constant to elementary/decode
